…ing an Inferrer

`Run` creates a new `PropertyName` each time `RunCreateOnce` only once than does the inferrence a whole bunch of times.

Before 

```
        Method |         Mean |     Error |     StdDev |      Gen 0 |    Allocated |
-------------- |-------------:|----------:|-----------:|-----------:|-------------:|
           Run | 1,001.777 ms | 77.603 ms | 20.1571 ms | 62000.0000 | 192203.37 KB |
 RunCreateOnce |     8.501 ms |  1.670 ms |  0.4338 ms |          - |      4.57 KB |
```

After

```
        Method |        Mean |      Error |     StdDev |      Gen 0 |    Allocated |
-------------- |------------:|-----------:|-----------:|-----------:|-------------:|
           Run | 1,147.55 ms | 172.197 ms | 44.7275 ms | 62000.0000 | 192203.37 KB |
 RunCreateOnce |    10.58 ms |   3.117 ms |  0.8096 ms |          - |      4.57 KB |
```

Since the caches are per `ConnectionSettings` and not per `Inferrer` this has no noticable performance boost nor implications for `5.x` but is still cleaner.