@russcam 2018-01-15
pulled this in @Mpdreamz