@russcam 2018-01-15
backported to 5.x (reinstated exception throwing in `string IUrlParameter.GetString(IConnectionConfigurationValues settings)` on `PropertyName`